refactor activities page by creating new subcomponent
